Update version to 0.12.2 and introduce new GateRegistry and LanguageConfig features
- Bump project version from 0.12.1 to 0.12.2 in CMakeLists.txt.
- Add GateRegistry and LanguageConfig classes for enhanced routing and language handling.
- Implement methods for managing page and route gates, including support for wildcard matching.
- Introduce obfuscation settings management with exclusion and preservation features.
- Add new utility functions for language prefix handling and localization in routes.
- Update ServerData to integrate new language and gate functionalities.
- Include unit tests for new features and ensure compatibility with existing functionality.
